+++ src/app/layout.tsx
@@ -1,10 +1,11 @@
 import type { Metadata } from "next";
 
-import { LayoutMain } from "@/components/LayoutMain";
+import { LayoutMain } from "@/components/layout/LayoutMain";
 
 import "@stellar/design-system/build/styles.min.css";
 import "@/styles/globals.scss";
 
+// TODO: update metadata
 export const metadata: Metadata = {
   title: "Laboratory - Stellar",
   description: "",
