+++ src/app/not-found.tsx
@@ -3,13 +3,15 @@
 import Link from "next/link";
 
 import { Routes } from "@/constants/routes";
+import { LayoutContentContainer } from "@/components/layout/LayoutContentContainer";
 
+// TODO: update 404
 export default function NotFound() {
   return (
-    <div>
+    <LayoutContentContainer>
       <h2>Not Found</h2>
       <p>Could not find requested resource</p>
       <Link href={Routes.ROOT}>Return Home</Link>
-    </div>
+    </LayoutContentContainer>
   );
 }
