+++ src/components/SdsLink.tsx
@@ -1,8 +1,8 @@
 import NextLink from "next/link";
 import { Props as LinkProps, Link } from "@stellar/design-system";
 
-/** Use `NavLink` instead of `Link` from Stellar Design System to support client-side routing. `NavLink` uses `Link` from `next/link` internally. */
-export const NavLink = (props: LinkProps) => {
+/** Use `SdsLink` instead of `Link` from Stellar Design System to support client-side routing. `SdsLink` uses `next/link` internally. */
+export const SdsLink = (props: LinkProps) => {
   return (
     <Link {...props} customAnchor={<NextLink href={props.href || ""} />}>
       {props.children}
