+++ src/components/layout/LayoutContentContainer.tsx
@@ -1,0 +1,9 @@
+export const LayoutContentContainer = ({
+  children,
+}: {
+  children: React.ReactNode;
+}) => (
+  <div className="LabLayout__container">
+    <div className="LabLayout__content">{children}</div>
+  </div>
+);
