+++ src/components/layout/LayoutWithSidebar.tsx
@@ -2,6 +2,6 @@
 
 import { ReactNode } from "react";
 
-export const LayoutSidebar = ({ children }: { children: ReactNode }) => {
+export const LayoutWithSidebar = ({ children }: { children: ReactNode }) => {
   return <div className="LabLayout__withSidebar">{children}</div>;
 };
